requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
cffi==1.17.1
contextlib2==21.6.0
cryptography==44.0.2
dj-ninja-auth==0.1.10
Django==5.2
django-ninja==1.4.0
django-ninja-extra==0.22.9
dnspython==2.7.0
email_validator==2.2.0
idna==3.10
injector==0.22.0
ninja-schema==0.14.1
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.10.1
python-dotenv==1.1.0
sqlparse==0.5.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
